backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-multipart
firebase-admin
requests
python-dotenv

